Refactor/add middleware module (#82)
* refactor: move middleware to a module

* fix: add middleware as subpath

* feat: change name module do middlewares
