+++ faster_sam/middlewares/remove_path.py
